Changes after discussion with Gemma
